distributed-data-system/requirements.txt
confluent-kafka==2.3.0
elasticsearch==7.17.9
pymongo==4.6.0
cassandra-driver==3.28.0
boto3==1.34.0
python-dotenv==1.0.0
Pillow==10.1.0
faker==20.1.0
numpy==1.24.3
opencv-python-headless==4.8.1.78
requests==2.31.0
python-dateutil==2.8.2
flask==3.0.0
prometheus-client==0.19.0
tenacity==8.2.3
ratelimit==2.2.1
gunicorn==21.2.0
werkzeug==3.0.1
